requirements.txt
asgi-redis==1.4.3
channels==2.0.2
Django==2.0.2
djangorestframework==3.7.7
Pillow==5.0.0